Eweasel tests for SCOOP exception handling.
The tests are written against the new semantics
by Benjamin Morandi, and currently they all fail.
Initially the tests were part of a naster thesis 
at ETH by Florian Besser.

The patch also includes a test for a bug
(test#scoop041) where the behaviour in molten
and frozen code differs when evaluating a
precondition on a controlled object.




git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@96579 8089f293-4706-0410-a29e-feb5c42a2edf